trying to fix things
