+++ libs/game-opt/sheet-ui/src/types/conditional.ts
@@ -1,6 +1,7 @@
 import type {
   Calculator,
   IConditionalData,
+  Tag,
 } from '@genshin-optimizer/game-opt/engine'
 import type { ReactNode } from 'react'
 import type { Field } from './field'
@@ -8,8 +9,12 @@ import type { Header } from './header'
 
 export type Conditional = {
   metadata: IConditionalData
-  label: ReactNode | ((calc: Calculator, value: number) => ReactNode)
-  badge?: ReactNode | ((calc: Calculator, value: number) => ReactNode)
+  label:
+    | ReactNode
+    | ((calc: Calculator<Tag, 'floor'>, value: number) => ReactNode)
+  badge?:
+    | ReactNode
+    | ((calc: Calculator<Tag, 'floor'>, value: number) => ReactNode)
   header?: Header
   fields?: Field[]
   targeted?: boolean
